Applied OOP principles
When given a specification for a feature or a new application, start by identifying the classes from the requirements or specification. Object-oriented programming represents the entities and concepts of an application as a set of classes. Each class has properties that define the data each object will manage. Each class has methods, which are the actions and behaviors that each object can perform. Then analyze the classes you identified and separate responsibilities as needed. Minimize coupling by ensuring each class has a single purpose. Maximize cohesion by reviewing the properties and methods of each class to confirm each one belongs to that class. Single-focus classes simplify maintenance and improve testability. No class is an island, so the next step is to look at relationships. The relationships between the classes define how the objects created from those classes work together to perform the operations of the application. Leverage reuse wherever possible. The power of object-oriented programming lies in its promise of reuse. By extracting commonality among a set of classes into a separate class, you have more reusable code. Extensive reuse of existing proven classes not only shortens development time, it also leads to more robust applications. There are several ways to achieve reuse, including extracting common functionality into a base class using inheritance. Reuse through a component library of general purpose classes and reuse through interfaces. This course also identified the four pillars, or characteristics, of object-oriented programming; abstraction, encapsulation, inheritance, and polymorphism. Abstraction describes an entity in simple terms, ignoring the irrelevant details. It reduces complexity by focusing only on what is important for the purpose of this specific application. Encapsulation allows for hiding the data and the implementation within the class. The data is stored in backing fields and accessible to the rest of the application through property getters and setters. The code is only accessed through the class interface. Inheritance allows derived, or child, classes to reuse all the code in the base or parent class. We created an EntityBase class that contained a standard set of properties and a Validate method. Each entity class then inherited from this EntityBase class to reuse its functionality. But in some cases, we don't want to reuse the base class functionality, so we leverage inheritance-based polymorphism by overriding the method. The idea behind inheritance-based polymorphism is that a base class can define a method and any derived class can override that method to provide its own definition and implementation, basically, providing its own shape for the method. Another way to leverage polymorphism is with interfaces. The idea behind interface-based polymorphism is that an interface can define a method, and any class can implement that interface to define its own code for that method, basically providing its own shape for the method. Generalized code can then access the appropriate shape based on the type of object that called it. Let's finish up this module with a look at next steps.

Next Steps